Win VISTA and 8 don'y allow selector changes even for 32 bit apps under wow64, doh! So moving to SFI under win64, selector work will need to be done for release on a 32 bit osk Compiles and gatekeper and simple client now compile and run at native 64 bit apps Also updated to VC11 compiler (few ices and stnadrd.h is a bust, likely to even more stripped down in future)

+++ src/libs/core/kdtree.cpp
@@ -86,12 +86,12 @@ void KDTree::buildSubtree( KDTreeNode* node, FaceArray& faces, AABB const& bound
 	// fill in the node as either a leaf or parent
 	if( faces.size() < g_minFacesPerLeaf || m_currentDepth == g_maxTreeDepth ) {
 		// push statistics
-		m_leafTriangleCount += faces.size();
+		m_leafTriangleCount += (unsigned int) faces.size();
 		m_leafDepthCount += m_currentDepth;
 		++m_leafCount;
 
 		// leaf node
-		node->setLeaf( faces.size() );
+		node->setLeaf( (unsigned int)faces.size() );
 		memcpy( node->getFaces(), &faces[0], faces.size()*sizeof( unsigned int ) );
 
 		// clear up our data
